+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## Release Notes:
 
-###vNext
+### vNext
 
 - Moved static base URL in the ENV config
+- Update Cere Wallet SDK to add CERE ERC20 tokens support
 
 ### v0.1.2