+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Moved static base URL in the ENV config
 - Update Cere Wallet SDK to add CERE ERC20 tokens support
 - Log and store session events
+- Fix wallet token reward amount
 
 ### v0.1.2