+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Update Cere Wallet SDK to add CERE ERC20 tokens support
 - Log and store session events
 - Fix wallet token reward amount
+- Fix bug with two leaderboards
 
 ### v0.1.2